requirements.txt
python-dotenv
ipykernel
groq
langchain
langchain_community
openai
flask